requirements.txt
requests==2.25.1
beautifulsoup4==4.9.3
openai==1.35.10
google-api-python-client==2.26.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.6
python-dotenv==0.19.2